refactor: 렌더링과 Rest 분리
