+++ src/main/java/roomescape/controller/AdminController.java
@@ -5,43 +5,23 @@ import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.request.AdminReservationRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
-@Controller
+@RestController
 @RequestMapping("/admin")
 public class AdminController {
     private final ReservationService reservationService;
 
     public AdminController(ReservationService reservationService) {
         this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    public String home() {
-        return "admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/time")
-    public String time() {
-        return "admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String theme() {
-        return "admin/theme";
     }
 
     @PostMapping("/reservations")
